requirements-cpu.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cpuonly=2.0=0
cryptography=38.0.4=py39h9ce1e76_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
my-package=1.1.dev2+g75a2c6f.d20230224=dev_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
omegaconf=2.3.0=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
pillow=9.3.0=py39h6a678d5_2
pip=22.3.1=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.0=hdb3f193_2
python-wget=3.2=py_0
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pyyaml=6.0=py39hb9d737c_4
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h5082296_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.13.1=py39_cpu
torchvision=0.14.1=py39_cpu
typing_extensions=4.4.0=py39h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py39h06a4308_0
wheel=0.38.4=py39h06a4308_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7f98852_2
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
